pyproject.toml
[tool.poetry]
name = "doccano-client"
version = "1.1.1"
description = "A simple client for doccano API."
authors = ["Hironsan <hiroki.nakayama.py@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/doccano/doccano-client"
repository = "https://github.com/doccano/doccano-client"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
]
packages = [{include = "doccano_client"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
dataclasses-json = "^0.5.7"
spacy = { version = "^3.4.1", optional = true }
spacy-partial-tagger = { version = "^0.9.1", optional = true }
tqdm = { version = "^4.64.1", optional = true }
pydantic = "^1.9.2"
requests-toolbelt = "^0.9.1"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
isort = "^5.10.1"
mypy = "^0.971"
darglint = "^1.8.1"
black = "^22.8.0"
responses = "^0.21.0"
types-requests = "^2.28.10"
pytest = "^7.1.3"
pyproject-flake8 = "^5.0.4"
taskipy = "^1.10.3"
mkdocs = "^1.3.1"
mkdocs-material = "^8.5.3"
mkdocs-same-dir = "^0.1.1"
vcrpy = "^4.2.1"
mkdocstrings = {extras = ["python", "crystal"], version = "^0.19.0"}

[tool.poetry.extras]
spacy = ["spacy", "spacy-partial-tagger", "tqdm"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
docli = 'doccano_client.cli.commands:main'

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'

[tool.flake8]
max-line-length = 120
max-complexity = 18
ignore = "E203,E266,W503"
filename = "*.py"
exclude = ".git,__pycache__,docs,build,dist,doccano_client/beta"

[tool.isort]
profile = "black"
include_trailing_comma = true
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
ignore_missing_imports = true
show_error_codes = true
exclude = [
    'doccano_client/beta',
]

[tool.taskipy.tasks]
isort = "isort . -c"
flake8 = "pflake8 doccano_client"
black = "black --check ."
darglint = "darglint"
mypy = "mypy doccano_client"
test = "python -m pytest"
